requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0